Modified Login, Sign-Up, and Profile
The Login and Sign-Up functions have been implemented further through the use of Flask-Login. This allows the use of sessions to keep a current user logged-in. I also changed the password to be hashed using sha256 and it is even stored like that in the database. Lastly, profile will now show the logged-in user's username, email, and phone number. A few pages were also modified with login_required to ensure that there is a user logged in to view the page. A few modifications may need to be made as issues may arise; specifically one may arise with the 'Remember Me' button.