Updating & Fixing Nav Bar + About Page
Updated the Navigation Bar to correctly reference the logout function instead of referencing login; which caused logging out to fail. Also removed the Account Drop Down from the Navigation Bar until the User logs in to the site. The Account Drop Down was replaced with a Login Tab which is then replaced by the Account Drop Down Tab again once successfully logged in. The Navigation Bar was also added to the Login and Sign-Up pages to allow the user to easily get out of them if desired.

Lastly, I updated the About Page to include the description that Shakirah made during her proposal.