Added Space and Delete
I have added space and delete functionality. The all fingers need to be up and the thumb needs to be close to the hand. For space it needs to be the front of the hand doing the sign (palm toward camera) and for the delete it needs be the same move but done with the back of the hand.